Tweak Makefile to handle patch versions with reproducible revisions
This idea is very simple: instead of bumping manually the patch number of a
version - frequent and tedious - we calculate a reproducible revision number.
It will be the number of commits since the earliest X.Y.0 tag. The above
signals the starting of a new cycle.
ClojureScript implements this strategy and it is really effective so we are
going to try it in here as well since 0.2.0.
The major and minor numbers are in the Makefile but will be propagated to all
the necessary files (project.clj, package.json, pom.xml).