Refactor all the specs in their own namespace
It seems to be agreed upon Clojurians that a library should not influence how
client code uses specs. This patch isolates all the specs in their own namespace.

Tests now include the specs for instrumentation as well. Finally this patch
removes the fonda.runtime namespace, now the FondaRecord is in fonda.execute.