[Fix #24] Make short-circuiting for anomalies optional (#38)
A user of the library might not need to use anomalies but just need to throw on
exceptions. This patch makes the anomaly machinery optional.

When the :anomaly? config key is truthy we now enable the functionality.

The patch is *Breaking* as it also changes the signature of fonda.core/execute
for a more node-friendly callback style - on-exception first then on-success.