Drop keyword only constraint for :name in execute specs and refactor
A couple of changes here. First of all the name specs in execute and core are
now the same and both allow keyword and string for :name keys.
Secondly, we refactor the specs so that including fonda.core.specs actually
adds everything to the registry.
Finally we fix the tests and issues that were detected with the change.